move random engine from engine.rs
